requirements.txt
numpy
streamlit
folium
streamlit-folium
pandas
geopy
sqlalchemy
streamlit-geolocation
Pillow
psycopg2-binary 
python-dotenv
numpy